chore(integrations/messenger): Refactor integration file structure (#13994)
