+++ integrations/messenger/src/webhook/index.ts
@@ -1,0 +1,2 @@
+import handler from './handler'
+export { handler }
